requirements.txt
streamlit
sqlalchemy
python-docx
pypdf
jinja2
pydantic
bcrypt
python-dotenv
